Remove duplicate element_bbox; reuse single implementation in viewport
Extracts element_bbox as a pub function in viewport.rs and removes the
duplicate points_bbox/element_bbox from spatial.rs. Rewrites
compute_bounds to delegate to element_bbox.
